requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
anthropic==0.62.0
python-dotenv==1.1.1
jinja2==3.1.6
